moved get_input_parameter to RyeReader
