+++ src/oxDNA_analysis_tools/duplex_finder.py
@@ -7,8 +7,7 @@ from collections import namedtuple
 import oxpy
 from oxDNA_analysis_tools.UTILS import geom
 from oxDNA_analysis_tools.UTILS.oat_multiprocesser import oat_multiprocesser
-from oxDNA_analysis_tools.UTILS.readers import get_input_parameter
-from oxDNA_analysis_tools.UTILS.RyeReader import describe, strand_describe
+from oxDNA_analysis_tools.UTILS.RyeReader import describe, strand_describe, get_input_parameter
 
 import time
 start_time = time.time()
